Create README.md
## WutMyGainz
An app inspired by the crypto craze that happens every 4 years or so. The purpose of the app is to easily calculate the user's gains/losses in the market.

## How It Works
It takes in to consideration your invested amount, your coin pairs to USD, and date purchased. The app then returns to you the percentage of gain/loss and well as how much the value of your "investment" increased or decreased by. It retrieves the spot price and historic price from the Coinbase API using Retrofit.

## Tech Stack & Libraries
- Retrofit
- Gson
- MVVM Architecture
- Room Database
- Live Data
- RecyclerView
- Coroutines
- ViewModel
- ViewPager